+++ song-tracker-backend/src/main/java/com/personalProjects/songTrackerBackend/repository/SongRepository.java
@@ -3,6 +3,6 @@ package com.personalProjects.songTrackerBackend.repository;
 import com.personalProjects.songTrackerBackend.model.Song;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SongRepository extends JpaRepository<Song, Integer> {
+public interface SongRepository extends JpaRepository<Song, Long> {
 
 }
